refactor: analytics sdk import
