+++ packages/sdk-v2/.ctiignore
@@ -1,5 +1,5 @@
 {
   "src/hooks/useAppRestart.api.native.ts": "*",
   "src/hooks/useAppRestart.api.ts": "*",
-  "src/sdk/analytics/**/*": "*"
+  "src/sdk/analytics/**/*.native.ts": "*"
 }
