+++ packages/sdk-v2/src/sdk/analytics/google/GoogleAnalytics.ts
@@ -3,13 +3,13 @@ import { omit, clone, defaults } from "lodash";
 import { IAbstractProvider, IAnalyticsProvider, IAppProps } from "../types";
 import { getUserProps } from "../utils";
 import apiFactory from "./api";
-import { IGoogleConfig, defaultConfig, IGoogleAPI } from "./types";
+import { IGoogleConfig, defaultGoogleConfig, IGoogleAPI } from "./types";
 
 export class GoogleAnalytics implements IAbstractProvider, IAnalyticsProvider {
   private api: IGoogleAPI | null;
 
   constructor(config: IGoogleConfig) {
-    const mergedCfg = defaults(clone(config), defaultConfig);
+    const mergedCfg = defaults(clone(config), defaultGoogleConfig);
 
     this.api = apiFactory(mergedCfg);
   }
@@ -30,7 +30,7 @@ export class GoogleAnalytics implements IAbstractProvider, IAnalyticsProvider {
     const { id, extra } = getUserProps(identifier, email, props);
 
     if (!api) {
-      throw new Error('GoogleAnalytics not initialized!');
+      throw new Error("GoogleAnalytics not initialized!");
     }
 
     api.setUserId(id);
@@ -41,9 +41,9 @@ export class GoogleAnalytics implements IAbstractProvider, IAnalyticsProvider {
     const { api } = this;
 
     if (!api) {
-      throw new Error('GoogleAnalytics not initialized!');
+      throw new Error("GoogleAnalytics not initialized!");
     }
-    
+
     api.logEvent(event, data);
   }
 }
