+++ packages/sdk-v2/src/sdk/analytics/google/types.ts
@@ -11,6 +11,6 @@ export interface IGoogleAPI {
   logEvent(event: string, data?: any): void;
 }
 
-export const defaultConfig: Partial<IGoogleConfig> = {
+export const defaultGoogleConfig: Partial<IGoogleConfig> = {
   userProperty: "user"
 };
