+++ packages/sdk-v2/src/sdk/analytics/index.ts
@@ -1,8 +1,10 @@
-export * from "./react";
-export * from "./sdk";
-export * from "./types";
-export * from "./utils";
 export * from "./amplitude";
 export * from "./google";
 export * from "./indicative";
+export * from "./mixpanel";
+export * from "./posthog";
+export * from "./react";
+export * from "./sdk";
 export * from "./sentry";
+export * from "./types";
+export * from "./utils";
