+++ packages/sdk-v2/src/sdk/analytics/sentry/Sentry.ts
@@ -1,19 +1,17 @@
-import { get, omit, forOwn } from 'lodash'
+import { get, omit, forOwn } from "lodash";
 
-import { IAbstractProvider, IAppProps, IMonitoringProvider } from '../types'
-import { getUserProps } from '../utils';
-import api from './api'
-import { ISentryConfig } from './types'
+import { IAbstractProvider, IAppProps, IMonitoringProvider } from "../types";
+import { getUserProps } from "../utils";
+import { SentryWeb as api } from "./api";
+import { ISentryConfig } from "./types";
 
 export class Sentry implements IAbstractProvider, IMonitoringProvider {
-  constructor(
-    private config: ISentryConfig
-  ) {}
+  constructor(private config: ISentryConfig) {}
 
   async initialize(appProps: IAppProps): Promise<boolean> {
     const { sentryDSN } = this.config;
-    const { env, version, ...tags } = omit(appProps, 'osVersion', '$once')
-    const sentryScope = { appVersion: version, ...tags }
+    const { env, version, ...tags } = omit(appProps, "osVersion", "$once");
+    const sentryScope = { appVersion: version, ...tags };
 
     if (!sentryDSN) {
       return false;
@@ -21,13 +19,11 @@ export class Sentry implements IAbstractProvider, IMonitoringProvider {
 
     api.init({
       dsn: sentryDSN,
-      environment: env,
-    })
-
+      environment: env
+    });
 
     api.configureScope((scope: any) =>
-      forOwn(sentryScope, (value: string, property: string) =>
-        scope.setTag(property, value))
+      forOwn(sentryScope, (value: string, property: string) => scope.setTag(property, value))
     );
 
     return true;
@@ -37,25 +33,25 @@ export class Sentry implements IAbstractProvider, IMonitoringProvider {
     const { id, extra } = getUserProps(identifier, email, props);
 
     api.configureScope((scope: any) => {
-      const user = get(scope, '_user', {})
+      const user = get(scope, "_user", {});
 
-      scope.setUser({ ...user, id, email, ...extra })
-    })
+      scope.setUser({ ...user, id, email, ...extra });
+    });
   }
 
   capture(exception: Error, fingerprint?: string[], tags?: object, extra?: object): void {
     api.configureScope((scope: any) => {
       // set extra
-      forOwn(extra || {}, (value: any, key: string) => scope.setExtra(key, value))
+      forOwn(extra || {}, (value: any, key: string) => scope.setExtra(key, value));
 
       // set tags
-      forOwn(tags || {}, (value: any, key: string) => scope.setTag(key, value))
+      forOwn(tags || {}, (value: any, key: string) => scope.setTag(key, value));
 
       if (fingerprint) {
-        scope.setFingerprint(fingerprint)
+        scope.setFingerprint(fingerprint);
       }
 
-      api.captureException(exception)
-    })
+      api.captureException(exception);
+    });
   }
 }
