+++ packages/sdk-v2/src/sdk/analytics/sentry/api.ts
@@ -1,3 +1,1 @@
-import * as SentryWeb from '@sentry/browser'
-
-export default SentryWeb
+export * as SentryWeb from "@sentry/browser";
