+++ packages/sdk-v2/src/sdk/index.ts
@@ -1,8 +1,9 @@
-export * from './base';
-export * from './claim';
-export * from './constants';
-export * from './faucet';
-export * from './microbridge';
-export * from './onboard';
-export * from './savings';
-export * from './storage';
+export * from "./analytics";
+export * from "./base";
+export * from "./claim";
+export * from "./constants";
+export * from "./faucet";
+export * from "./microbridge";
+export * from "./onboard";
+export * from "./savings";
+export * from "./storage";
